+++ lib/src/main/java/io/ably/lib/objects/LiveObjectsAdapter.java
@@ -30,6 +30,6 @@ public interface LiveObjectsAdapter {
      *
      * @return the maximum message size limit in bytes.
      */
-    long maxMessageSizeLimit();
+    int maxMessageSizeLimit();
 }
