+++ lib/src/main/java/io/ably/lib/transport/Defaults.java
@@ -53,7 +53,7 @@ public class Defaults {
     /* CD2h (but no default in the spec) */
     public static long maxIdleInterval = 20000L;
     // 64kB, as per CD2c
-    public static long maxMessageSize = 65536L;
+    public static int maxMessageSize = 65536;
     /* DF1a */
     public static long connectionStateTtl = 120000L;
